bitn fixes
