+++ Dragon's Trawel/src/mainGame.java
@@ -13,7 +13,7 @@ import java.util.Scanner;
  * calls all the other classes, and holds the defining gameplay types and inputs.
  */
 public class mainGame {
-	public static final boolean bumpEnabled = false;//a boolean for an update
+	public static final boolean bumpEnabled = true;//a boolean for an update
 
 	//instance variables
 	public static Scanner scanner = new Scanner(System.in);
@@ -144,6 +144,7 @@ public class mainGame {
 		new Oracle().load();
 		new TauntsFactory();
 		new BookFactory();
+		new BumperFactory();
 		story = new StoryNone();
 		extra.println("Dragon's Trawel v0.5.1");
 		extra.println(
